pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=2.1.9",
    "langchain-google-vertexai>=2.0.28",
    "langchain[google]>=0.3.27",
    "langgraph>=0.6.5",
    "python-dotenv>=1.1.1",
]
